Temporarily disable type sort in MDZipView and move code to header.
These changes are required to make refactoring slightly quicker. These
features should be reintroduced later.
