+++ test/md_zip_view_test.cpp
@@ -549,48 +549,3 @@ TEST(MDZipView, derived_standard_deviation) {
   ++it;
   EXPECT_TRUE(std::isnan(it->get<Data::StdDev>()));
 }
-
-TEST(MDZipView, type_sorting) {
-  Dataset data;
-  data.insert(Coord::X{}, {}, 1);
-  data.insert(Coord::Y{}, {}, 1);
-  MDZipView<Coord::X, Coord::Y> a(data);
-  MDZipView<Coord::Y, Coord::X> b(data);
-  MDZipView<Coord::Y, const Coord::X> b_const(data);
-  EXPECT_EQ(typeid(decltype(a)), typeid(decltype(b)));
-  EXPECT_NE(typeid(decltype(a)), typeid(decltype(b_const)));
-}
-
-TEST(MDZipView, type_sorting_nested) {
-  Dataset data;
-  data.insert(Coord::X{}, {}, 1);
-  data.insert(Coord::Y{}, {}, 1);
-  MDZipView<Coord::X, MDZipView<Coord::Y>> a(data);
-  MDZipView<MDZipView<Coord::Y>, Coord::X> b(data);
-  EXPECT_EQ(
-      typeid(decltype(a)),
-      typeid(
-          MDZipViewImpl<Dataset, Coord::X, MDZipViewImpl<Dataset, Coord::Y>>));
-  EXPECT_EQ(typeid(decltype(a)), typeid(decltype(b)));
-}
-
-TEST(MDZipView, type_sorting_two_nested) {
-  Dataset data;
-  data.insert(Coord::X{}, {}, 1);
-  data.insert(Coord::Y{}, {}, 1);
-  data.insert(Coord::Z{}, {}, 1);
-  MDZipView<Coord::X, MDZipView<Coord::Y, Coord::Z>> a(data);
-  MDZipView<Coord::X, MDZipView<Coord::Z, Coord::Y>> b(data);
-  MDZipView<MDZipView<Coord::Y, Coord::Z>, Coord::X> c(data);
-  MDZipView<MDZipView<Coord::Z, Coord::Y>, Coord::X> d(data);
-  EXPECT_EQ(typeid(decltype(a)),
-            typeid(MDZipViewImpl<Dataset, Coord::X,
-                                 MDZipViewImpl<Dataset, Coord::Y, Coord::Z>>));
-  EXPECT_EQ(typeid(decltype(a)), typeid(decltype(b)));
-  EXPECT_EQ(typeid(decltype(a)), typeid(decltype(c)));
-  EXPECT_EQ(typeid(decltype(a)), typeid(decltype(d)));
-  MDZipView<Coord::X, MDZipView<Coord::Y, Coord::Z>> a_const(data);
-  EXPECT_EQ(typeid(decltype(a_const)),
-            typeid(MDZipViewImpl<Dataset, Coord::X,
-                                 MDZipViewImpl<Dataset, Coord::Y, Coord::Z>>));
-}
